Use people picker for owner criterion
Part of story #27191 ease search on document owner

When we search for documents owned by a user but we don't know exactly
their id or their username, then we are stuck. This contribution lets
the user choose a user via a user autocompleter.
Since there is the autocompleter, the popover helper becomes outdated and
not needed anymore.

Only document owner are retruned by the select, no matter if the user is
in active, suspended or destroyed state

Change-Id: I9603eeb46ad347d6fdf163c740d0eea98e9c2259