Stop using `jest-jasmine2` to run the tests of packages using `angular-mocks`
`angular-mocks` is only compatible with Jasmine and Mocha since it has
not been updated for a while. As a result we still have have to use
jest-jasmine2 instead of the default jest-circus.
In fact `angular-mocks` can run inside another test runners as long as
you expose compatible APIs and you make it thing it is executed with
Jasmine.

Part of request #27909: Move packages using `angular-mocks` to Vitest

Change-Id: Ic909edfa7df39d8e08c488b8bb7ae4bebaef319c