fix: request #27903 delete field value should not be possible when it's use on transition
Given a tracker with a static list field, when we define a workflow based on this field,
all values can be used to create transitions.
When having a value that has only transitions "from" but never "to",
then this value can wrongly be deleted (if all other conditions are met).

This leads to have transitions from deleted value yet defined for the workflow,
which breaks the UI by wrongly displaying these transitions.

A forgupgrade (the same as previously defined in
202203221026_clean_workflow_transitions) is added to clean the database
from the transitions that have either the "from" or the "to" field value linked to a
deleted value.

Change-Id: I15327b0fab3d96f425a941ebc713a188bded3b9e