request #27911 Have too many artifact can make the tracker report very slow
If there are too many artifacts in a tracker (60k reported), access to the report can be very slow and eventually lead to a timeout.

Technical information :

This comes from a query that can take a very long time.
- we search all the matching ids of the current report
- We search for all field of those matching ids with IN (matching_ids)
  which can take too much time.

--> we decided limit the number of matching ids before search all the
fields informations

We can see for the aggregate functions to have better performances if needed.

Performances :
Tests has been done with 65k artifacts / 16 fields displayed in report

Before :
 - 9.23s
 - 8.02s I/O wait
 - 49,3 MB memory
 - 953 SQL queries in 7.32 s

After :
 - 5.33s
 - 4.13s I/O wait
 - 49,9 MB memory
 - 1055 SQL queries in 3.85 s

Change-Id: Ie3a9103526dedc978ffbfdfead63d24357db2f17