feat: Docker Postfix configuration uses the email_* configuration variables
This is part of story #27133 send mail without postfix

You can use TULEAP_EMAIL_TRANSPORT and TULEAP_EMAIL_RELAYHOST env variables
to define your mail transport used with Docker:

- TULEAP_EMAIL_TRANSPORT=smtp
- TULEAP_EMAIL_RELAYHOST=mailhog:1025

For backward compatibility, you can only provide TULEAP_EMAIL_RELAYHOST
and this will not lead to an invalid mail configuration.
In that case, Sendmail will be used and a warning message will be displayed to
the user.

Change-Id: I50b747ee779635f35c405f5de3542da506dcc250