Decouple RESTCurrentUserMiddleware from the controllers
This is needed because we will need to expose a different authentication
method to allow ONLYOFFICE to download documents. Decoupling the
existing way will allow us to re-use the same controller.

This is a technical changes, nothing is expected to be different from
the end-user point of view.

Part of story #27555: display Office documents stored in the Document plugin

Change-Id: I8beed63f39c3f324be4eb147c92cb59f2074d369