feat: Add Download in dropdown menu
Always display a "Download" item in the dropdown whenever the document
is a File. That way people easily know how to download a file, even if
the initial link on the title is captured by another plugin like OnlyOffice.

Only the dropdown in the tree view is concerned by this contribution,
quicklook dropdown has a dedicated download button. We will see later
how to deal with download and open in OnlyOffice combination.

Part of story #27555: display Office documents stored in the Document plugin

Change-Id: I6e7c1193c4bd215e9dfdbe24878d5c47a0f9817e