feat: Handle commits of a push sequentially
part of request #27896 Smart commit is not smart enough

How to test:
No functional change expected in the Git closing references feature:
- if you push two commits that have a "closing comment" for the same
artifact, they are handled in sequence instead of in parallel
- Tracker workflow manager will still comment twice on the artifact

Notes:
- Due to the existence of inline caches both in Tracker_ArtifactFactory
and Artifact::getStatus(), it does not solve the bug: the code will try
to close the artifact twice. Another contribution will be needed to
ensure we do make the SQL query twice to check whether the artifact is
open.
- It makes more sense to check that the artifact closure is allowed in
DefaultBranchPushProcessor. After all, it is where we produce events for
Tracker. Previously it was done at the beginning in PushCommitsAnalyzer.

Change-Id: Id3b3dfc48eebeafec44e72b1cafcec5646d27b1c