refactor: introduce {Header,Footer}Configuration
We should stop using associative array to configure display of
header/footer.

This focuses only on what is being used by OpenInOnlyOfficeController,
the remaining associative key are left untouched to keep the change
small.

Took the opportunity to remove Layout::footer() which seems to be dead
code.

Part of story #27555: display Office documents stored in the Document
plugin

Change-Id: I05ef0abc31823a1e168f347d098c946cda4bb281