Introduce the POST gitlab_group REST route
Part of story #26799 associate a Tuleap project and a GitLab group

That contribution introduce the REST route, for now, it only retrieves
and creates the Gitlab integration on the Tuleap side and nothing else.

The route is behind a feature flag, so you must enable it if you want to
see the route.
To enable it:  tuleap config-set feature_flag_gitlab_link_group 1

You can test this contribution in the API Explorer => POST /gitlab_groups. It returns the gitlab group id and the number of integration created in Tuleap

Note: Your token must have the `api` scope

Change-Id: I9bdfd9c9d09c787c818d7bc458596fd91e32948b