fix: Do not display encrypted secret with config-get
For secret stored in config, admin must use --reveal-secret switch for
`tuleap config-get` command in order to read the value.

However if admin does not know about this option, until now they will
get the encrypted secret. It may be confusing (is it the raw secret that
is stored?) and probably useless (how can I decrypt it to have the
value?). In order to mitigate those issues, the proposition is to raise
an error if admin tries to display a secret without --reveal-secret
switch.

Part of story #27555: display Office documents stored in the Document plugin

Change-Id: Ia16666f6b6a8b56824c924a21b83bb1e2d3b6afc