feat: send one big event to Tracker instead of small events
part of request #27896 Smart commit is not smart enough

No functional change expected in the Git closing references feature:
- if you push two commits that have a "closing comment" for the same
artifact, Tracker workflow manager will still comment twice on the
artifact.

Notes:
Instead of sending an event for each commit message, we send one big
event for the entire branch that was pushed. It contains all the commit
messages that we could read. If the code can't read one commit message,
it does not stop and tries to read the next one.
The single big event will allow us to deploy local caching to avoid
trying to close again an artifact that has been closed in the same push.

Change-Id: I69da9cc8d7e0f63f4e938806e446bfdb5c4b6d3f