feat: Do not throw an error if description changeset value is missing in definition
This closes request #28249 Do not throw an error 500 if description changeset value is missing in definition

If for some reason a changeset value for hardcoded description field is
missing in a test definition, this leads to an error 500 in the REST call and the
testmanagement campaign content cannot be displayed. This situation must
not lead to an error but be considered as an empty value instead.

In user interface, these kind of error messages are not understandable and usefull for
the user. So we do not display them anymore, only a generic error message is
displayed.

Change-Id: Ia0cd6769740e4111e12077f72210b5fabad4f64b