Start loading the ONLYOFFICE editor
At this stage the editor stays in its loading state since we do not
provide information on what it is supposed to load. This will be added
later.

The ONLYOFFICE editor is loaded on a sandboxed page independant from the
rest of Tuleap. The goals for doing this are multiple:
* we want to avoid side effects as much a possible: ONLYOFFICE docs API
  inject itself into `window` and we cannot guarantee it will always
  play well with the Tuleap codebase
* a compromission of the ONLYOFFICE document server should have a
  limited impact on Tuleap (i.e. it should only get access to documents
  users try to open)
* we do not want to pollute on own log file with issues coming from the
  ONLYOFFICE editor

Adjustments will be made to the sandbox once we can ask ONLYOFFICE to
load a document so we can see what work and what does not.

Part of story #27555: display Office documents stored in the Document plugin

Change-Id: Ia18b9bc948033ad3467ef38dfce4134f02428650