Closes request #27537: Update dev/build environment [W33 2022]
Main changes:
PHP: 8.0.18 -> 8.0.22
Composer: 2.3.5 -> 2.3.10
Nix: 2.8.0 -> 2.10.3
Cosign: 1.8.0 -> 1.10.1
Node: 18.0.0 -> 18.7.0
pnpm: 7.6.0 -> 7.9.3

The derivation building the `tuleap-git-bin` package has been updated to
disable the "install check" phase. Since [0] it is enabled when building
on compatible platforms but in our case it does not really work since we
set a different build prefix path. It is not really an issue either
since we want to check if the package work when used in the Tuleap
context so we are a bit forced to do the install checks ourselves when
upgrading Git.

[0] https://github.com/NixOS/nixpkgs/commit/41485e7337baca768dc542c553a9d66030df7b33

Change-Id: I31fc76e725b3ff5c1ff4a094e2264a3ab39d04ab