Drop onClick handler abuse made by Prototype.js in order to support IE and old Opera browsers
Prototype.js has some code to determine if it is running under IE or old
Opera to be able to read attributes on an element. This is not needed
anymore and the way it determines it by using the onClick does not play
nice with a CSP using a nonce for `script-src`.

No functional changes.

Part of request #17967: Deploy a useful content security policy

Change-Id: I8f9b178321d88eee1339a3531fafa50f6490cc9b