fix: None value must be added in required selectboxes without default value
This closes request #28255 None value must be added in required selectboxes without default value

Given a tracker that have a mandatory selectbox without any default value set,
when opening the artifact modal, then the selectbox must display "None" as the
selected value to force the user to select a value.

This is the current behavior in the Artfiact creation UI in tracker plugin.

Note that nothing must be done for multi selectbox, the current behavior
in the Artifact modal is the one expected.

How to test
-----------

* In tracker administration, add a non mandatory selectbox without a default
  value
* Open the artifact modal creation (in Agiledashbord for instance)

=> "None" must be selected as default value for the selectbox

* In tracker administration, update the selectbox to mark it mandatory
* Open the artifact modal creation (in Agiledashbord for instance)

=> "None" must be selected as default value for the selectbox

* In tracker administration, update the selectbox to add a default
  value
* Open the artifact modal creation (in Agiledashbord for instance)

=> The default value must be selected in the selectbox

Change-Id: I998b62bc3a4451066b91be1f92b2a0aa39bec46c