Give information to ONLYOFFICE to load documents
Note that the sandboxing have been relaxed to add `allow-same-origin`
because ONLYOFFICE want to access the local storage.
This means the sandboxing is now mostly an illusion. Tuleap can
be completly compromised in case of a sec issue in ONLYOFFICE or if an
untrusted value is set as the ONLYOFFICE document server URL.

Part of story #27555: display Office documents stored in the Document plugin

Change-Id: I34e28a7cbdf527e6651433fcf58658bb8b90fd73