Save the gitlab group in database
Part of story #26799 associate a Tuleap project and a GitLab group

Since the patch is already big enough, the link bettween integration and
group is not made yet, it will be done in a dedicated contribution

Pre-conditions:
 - Make sure to enable the feature flag
 - The Gitlab token used should have at least the "api" scope.

You can test via the API Explorer, if all it's ok, 201, the
id of the created group and the number of integrated repositories are
returned.
If the group already exists, then you should have a 400.

Change-Id: I9e2bd9ff37d5f525507e0b52f167c84cbabc0e57