feat: Switch onlyoffice plugin to project scope
Administrator may not want to deploy ONLYOFFICE integration to
the whole platform. To reduce costs or for progressive deployment, we
should be able to restrict access to the plugin.

Restriction is managed in plugins administration.

If a project is not allowed to access to ONLYOFFICE
=> Documents URLs are not overridden
=> Direct access to the OO document is rejected

Part of story #27555: display Office documents stored in the Document plugin

Change-Id: I18b28192ec9d6f8a9248c33b50e7402c0439565c