feat(jira): Auto create users from LDAP at Jira Import
When Jira <-> Tuleap user mapping is done, the email address is being
used. When the user email is not found and when LDAP plugin is used, we
try to find the user and to provision its account to keep data as close
as possible from source.

This commit also introduce a command for Tuleap developers to stuff the
LDAP server with users. It's for dev purpose only so implementation make
a lot of assumptions on what is enabled & all.

How to test:
1. Have a Jira project with a user (for instance assignee or sumbitter
   of an issue) whose email address doesn't exist in your tuleap.
2. Create this user in your development LDAP with
   tools/utils/tuleap-dev.php add-ldap-user login "Real name" email@example.com
3. Run import (either in tracker or whole project)
4. Profit.

Closes request #28265 Auto create users from LDAP at Jira Import

Change-Id: I8188912cd88d07328f636373801fa8bd1235e8ed