feat: Be able to select the board to import in Jira project import
This closes request #28267 Be able to choose the scrum board to import ar Jira project import

In Jira side, multiple scrum boards can be defined per project.
However, at Tuleap side, only one scrum agiledashboard can be defined.
For now, we import the first Jira scrum board found in project when doing a Jira project import.

To be able to fine-tune this import, a new option is added to provided a board id
to retrieve during the import: jira-board-id

Error are raised if either the provided board is not a scrum one,
or if the board is not in the Jira project we want to import.

Not found board are silently ignored.

Change-Id: Icdb1097286ad447a5d88d63c8ed227bbb0ca09bf