feat: Call REST endpoint from frontend
When `fts_db` is enabled, a REST call is made to retrieve results. For
now we only display the title, but it will evolve later to provide rich
content.

* the REST call is not made if keywords is less than 3 chars
* a spinner is displayed to show that something is happening
* If REST call is in error, an error message is displayed
* only the first page of results is displayed for now
* to not flood the server for every keystroke while user is typing, the REST call is delayed

Note: pagination will be done in a dedicated commit

When `fts_db` is disabled, no functional changes ("No results" is still
displayed when no projects and no recent items match the search).
Note: there is a first try to call the search API, but since we end up
in a 404, we stop searching stuff.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: I8ba3087c103db07b9ea2de3fa5410b095f996657