feat: do not wrap anymore with arrow keys
Open SwitchTo modal,
hit `tab` key to go to the project list with the keyboard.
If you hit up arrow key while the focus is on the first project, then
nothing happen.
If you hit down arrow key while the focus is on the last project, then
nothing happen.
The same for recent entry list.

We remove this behavior because we are not sure that it is useful and it
would have make keyboard navigation algorithm more difficult when in
search mode (Do we really want to switch the focus from the first
project to the last search results?).

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: Ib406196c145c7ba3b46c37e404b6b01289d75d62