Drop existing/Do not add DB entry when content to index is an empty/small string
Storing such entries is never going to yield interesting results so
storing them is a waste of space. Same for small strings.

Note the implementation has been managed directly into the DAO as it is
not really business logic. Some implementations might prefer having all
the updates to have the ability to detect trends and recommendations.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: Ieb2db72920ba50370827d39269c7936c889c0aa1