feat: focus to quick link with arrow right key
* Open SwitchTo modal
* Press `tab` until you reach a project you are admin
* Press arrow right key ➡️ admin quick link is focused
* Press `tab` until you reach a recent items with quick links
* Press arrow right key ➡️ the firt quick link is focused

Notes:
Only switch from project or recent item to quick link is implemented by
this contribution.
This temporarily breaks ability to switch from a project to recent items list with
arrow right. It will be completed later, to not clutter to much the
review.

Took the opportunity to align quick links implementation between a project
and a recent item: both have an array of quick link. Even if project has
only one link (for now), it is easier to have a single definition.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: Ib9b06ddd6f763d4a2f40dd79a59ea42d79ba4015