feat: Search in artifacts only
Inform the user that the search is only done in artifacts.

This is done via a question mark icon next to the Search results title.

Minimal effort is done for this since it is not supposed to stay (lol).
A more complete description would require to let plugins/backend
describe what is being supported (sic).

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: Ic9ee19bda365c40ae0421bb685dfecfe81d60dd9