refactor: Separate tests for repository integration
part of story #26799 associate a Tuleap project and a GitLab group

No functional change.

Notes: Interface names should not have "Dao" in them, at least not if we
want to separate business logic from Database. I therefore renamed it.
This also rewrites the test to have two cases: one where the
repositories of a group are already integrated, and another where they
are not. Having the stub change in each case makes the control flow
clearer than having a stub that sometimes returns true or false.

Change-Id: Ic759862776e8cc50d3d0913f4e68d2afe3b8355d