feat: No left/right navigation while searching
When user starts searching for something, the display of lists change
from an horizontal layout to a vertical layout. Therefore navigating
from a list to another by using left and right arrow keys becomes weird.

For this situation, we will be able to navigate from a list to another
by using up and down arrow keys, but it is the subject of a dedicated
changeset. #BabySteps

No functional changes in horizontal layout (aka no search).

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: I6fef0dad1329562ff163d0d48fda704bcbc93cee