fix: default lang for LDAP provisionned users
LDAP was relying on an old fashioned way to find the appropriate lang
for a user. By default, it should be the platform configured language
because we assume it's the lang that is the most commonly shared by the
users.

Should close request #28281 Approval mail to new user can have multiple languages

Change-Id: I4b262ecac2e4ba7e43fe4b94c618640adb658739