refactor: revert usage of postJSON in fulltext search
postJSON has been updated[0] so that it returns directly a decoded JSON
instead of a response.

This behavior is fine for most usages, but fulltext search needs to do
multiple queries according to the pagination size. Since this size is
only known by the Response we now switch from postJSON() to the new
post() method[1] which is providing what we will need.

No functional changes, this is a baby step to prepare the field to more
complex pagination management.

[O]: git #tuleap/stable/36cd013b13921389793d349d25b8edafac9447c8
[1]: git #tuleap/stable/d7f81c9b95592afb91cba911d89f4555f449f82c

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: Id14e32673d6ba97b4e096a455dce5f172013208e