refactor: rebuild the plugin_gitlab_group table
Part of story #26799 associate a Tuleap project and a GitLab group

How to test:
- run `make dev-forgeupgrade`. ⚠ If you had data in the
`plugin_gitlab_group`, it will be erased!
- Try the REST route with the API explorer. It should succeed if you
give it valid parameters. There should be a new row in
plugin_gitlab_group table.

Notes:
`project_id` is added so that we have a way to link the GitLab group to
the Tuleap project. Without it, it's hard to tell whether a given Tuleap
project is linked to any group or not. Given that the REST route is the
only way to fill this table, and the REST route is behind a feature
flag, I assumed that the only users would be developers. Thus, I did not
attempt to write a db migration, the basic tactic of "drop + create" is
used and data will be lost.

`prefix_branch_name` is renamed to `create_branch_prefix` to have
consistent naming with
`plugin_gitlab_repository_integration_create_branch_prefix` table.

Change-Id: I9ccbad04d2e1af199674538c8a3700d9b1cb7b50