Indexing all items does not rely on async processing
Indexing all items is an operation that generates a lot of events which
can saturate standard Tuleap operations.

Note this means the full-indexing is a bit slower since it does not
benefit from the parallelism provided by multiple workers. While this
can be an inconvenience it also acts as a rate limiter making it to run
it during standard operations if you have a bit of spare capacity.

To test, launch `tuleap full-text-search:index-all-items` it should
index everything without adding async worker events. If you edit an
artifact, the indexing is still managed asynchrounously.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: I4c609662a527ac768e6d4723f5a9e22aee4393c2