Make possible to resume the "full-indexing" of items
The full-indexing is now 2 distinct steps:
* one identify all the items that can be indexed
* another that do the actual indexing of the identified items

The identification of items to index is automatically done when the
plugin is enabled.

To test, you can play the command `full-text-search:identify-all-items-to-index`
then `full-text-search:index-all-pending-items` (you can stop and
restart it at will).

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: I3cde6678ae1fc22978b1e7fac8eadffb58b49a53