feat: Display 15 results
Display only 15 results. If results are spread accross multiple pages
then they all are retrieved until we have 15.
If there are more than 15 results the user can see, then a message is
displayed at the end of the results list so that the search query can be
refined.

There are some glitches:
* If there are a lot of item matching the query, but few the user can
see, then user has to wait until all parallel requests are finished
before seeing the results. => I suggest to display those results as soon
as they are available in a dedicated contribution.
* If user enter a search term, then change it, the initial bunch of
requests are still performed: 1) it consumes useless resources 2)
initial term results might be displayed instead of the new ones. =>
we have to find a way to cancel the pending requests (dedicated
contribution?)

Note: post() is updated to be able to pass limit/offset in POST query.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: If558c23f6d7c8af5dfaf835b47999604776b42dd