refactor: GroupCreatorTest should be a sociable test
Part of story #26799 associate a Tuleap project and a GitLab group

No functional change.

Interfaces are good but for high-level classes like GroupCreator, it's
nice to be able to test all its execution, including its dependencies.
It's "sort-of" an integration test, except all the API calls and DB are
stubbed. The rest of the logic stays the same. It can help check that we
did not break things when refactoring the inner dependencies.

Change-Id: I17508330e779f6348288982b7fc04d9101829b51