feat: Apply configuration on new integrations
Part of story #26799 associate a Tuleap project and a GitLab group

How to test:
The REST route now accepts configuration parameters, such as:
- whether it allows artifact closure or not (boolean)
- the branch prefix (string or null)

If you give this data to the route, it will apply it to all new
integrations. If a gitlab project was already integrated previously in
the project, its configuration is not modified.

If successful, it should respond with code 200 (and not 201, since we
don't provide a Location Header and the GET URI will not be the same).

If a branch prefix would result in an invalid Git branch name, it will
respond with code 400 Bad request.

Change-Id: I0d37ff0bad388b2fb8a216088defc205b97ee751