feat: filter projects and history by keywords
Instead of searching for the complete string, we should consider the
filter as a space-separated list of keywords.

This will make the filter behavior a little bit closer to what is being
done in search results (even if the latter depends on the backend
implementation).

Highlighting is now done per words.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

Change-Id: I3f5961e63f19f2d164128b020bd5e631da61e682