fix: user_navigation e2e tests is broken
Behavior of filter has been changed to better reflect what is being done
on the backend side: Searching "Explicit backlog" will return both
"Explicit backlog" and "Implicit backlog" since they contain
"backlog"[0].

Cypress is not happy about this change because when we want to click on
an admin icon, it finds two occurences, and don't know which one to
click on. Since the behavior we are testing is to reach the project
administration, we now select the first one.

Part of story #26771: Index artifact fields (REST + Switch to) on updated artifacts only

[0] git #tuleap/stable/a0fc0759742a713ad66f4f063fbbc3ffe2d196c2

Change-Id: Ibff1a2857b899e19ef80b79bafa4b5e89ba928ac