refactor: move File field types to Domain
Part of request #28830 replace File upload streams by TUS client

No functional change.

Notes:
Those types are sort of "the Domain" of the File field already. We
should try to limit the dependency on the REST API representations,
but that will be a task for later. The rename will help reuse those
types for the file upload.

Change-Id: I90ad0a5c5e01c4063ac781d96b90c237ad43bb54