feat: Allow passing a Description for a File
Part of request #28830 replace File upload streams by TUS client

How to test:
- If you provide a Description for a File in POST
/tracker_fields/:id/files, it will be saved in the `tracker_fileinfo`
DB table.

It is not provided by existing callers so it is optional, to avoid
causing a breaking change. If null, it is defaulted to an empty string,
so the same behaviour as before is preserved.

No functional change on image drag/drop in Trackers.

Change-Id: I9ef134065c34ab25f2b632f911e91fc953c13901