Make the token used to authenticate the Jenkins plugin hook trigger usable only once
No functional changes expected, the token is opaque for the Jenkins
Tuleap plugin. The way it is generated and managed on the Tuleap side
does not affect it.

The initial implementation made in
68785d2042ff30474bc336c14d682d362adaccc6 was using authenticated
encryption to make sure the provided token was valid. It was made this
way to avoid creating a new DB table to make backporting the fix accross
multiple Tuleap versions seamless.

This contribution replaces this initial approach for the same approach
used in other places in the Tuleap codebase. This comes with the
following advantages:
* less cryptography (we only use hashing to store the token securely in
  the DB)
* the token can be used only once

Closes request #28257: Missing authentication of webhook requests made to the Jenkins Branch Source plugin

Change-Id: I5776c25428615859b0ef4e0e32d2476339422029