feat: Replace File upload streams by TUS client
Part of request #28830 replace File upload streams by TUS client

Change expected: the TUS API for Trackers should be called instead
of the Artifact temporary files API. Otherwise, no expected functional
change:
- when you attach several files to several File Attachment fields,
they are all uploaded in parallel when you submit the modal. When the
uploads are finished, the create / edit REST API is called and the files
are attached.

Notes:
Error handling still leaves to be desired: you need to scroll up to see
error messages, like before. It can be improved after redoing the
"submit" handling, currently AngularJS promises are a major hindrance.
Additionally, error messages are still very technical and untranslated
most of the time. For example, exceeding the max upload size displays
the message from the backend, which shows file sizes in bytes (instead
of MiB).

The modal bundle weight is reduced: 647 KiB -> 615 KiB.
It is not as reduced as I hoped for. We could have some gains by
deduplicating "tus-js-client" which is included a second time as part of
the CKEditor image upload lib, but that will require a version bump at
least.

Change-Id: I552e194d2129fd401935394f67c61c5891f4ee80