Init `fts_meilisearch` plugin and `tuleap-meilisearch-server` package
To test the `tuleap-meilisearch-server` package you can build it with
`nix-build ./plugins/fts_meilisearch/additional-packages/tuleap-meilisearch-server.nix`,
copy the resulting inside your dev env and install it. You should be
able to execute `/usr/bin/tuleap-meilisearch -V`.

Meilisearch is built statically (like the other binaries we deploy) to
avoid pain with incompatible glibc versions or even other libs.
Meilisearch is mainly written in Rust but some of the dependencies it
relies on links against C code. To make our (sort of) cross-compilation
easy and not battle against Musl and gcc we use Zig which make the process
seamless [0].

Part of story #28835: have a full-text search backend based on Meilisearch

[0] https://andrewkelley.me/post/zig-cc-powerful-drop-in-replacement-gcc-clang.html

Change-Id: I781acc50c6aa9fa3a573358ba32e4bd9470a0047